+++ zcash_primitives/src/sapling/prover.rs
@@ -44,6 +44,32 @@ pub trait SpendProver {
     ) -> Self::Proof;
 }
 
+/// Interface for creating Sapling Output proofs.
+pub trait OutputProver {
+    /// The proof type created by this prover.
+    type Proof;
+
+    /// Prepares an instance of the Sapling Output circuit for the given inputs.
+    ///
+    /// Returns `None` if `diversifier` is not a valid Sapling diversifier.
+    fn prepare_circuit(
+        esk: jubjub::Fr,
+        payment_address: PaymentAddress,
+        rcm: jubjub::Fr,
+        value: NoteValue,
+        rcv: ValueCommitTrapdoor,
+    ) -> sapling::circuit::Output;
+
+    /// Create the proof for a Sapling [`OutputDescription`].
+    ///
+    /// [`OutputDescription`]: crate::transaction::components::OutputDescription
+    fn create_proof<R: RngCore>(
+        &self,
+        circuit: sapling::circuit::Output,
+        rng: &mut R,
+    ) -> Self::Proof;
+}
+
 /// Interface for creating zero-knowledge proofs for shielded transactions.
 pub trait TxProver {
     /// Type for persisting any necessary context across multiple Sapling proofs.